development: created new method that creates a SplitBoundary between regions; method does not work fully yet chore: rewrote method that identifies InterFace neighbors using NodeManifold rather than Node pointers chore: updated some member functions of MeshManager

+++ source/includes/model/CSMP_global_enumerations.h
@@ -14,9 +14,9 @@ CSMP global constants and enumerations
 ========================================
 */
 
-/// initialisation of const to maximum value that size_t can take
+/// initialisation of const to maximum value that uint32_t  can take
 const size_t NULL_IDX(std::numeric_limits<uint32_t>::max());
-const short  UNSPECIFIED(-1);
+const int    UNSPECIFIED(-1);
 
 /**
 @defgroup CSMP_global_enumerations
